requirements.txt
playwright==1.58.0
pytest-playwright==0.7.2
pytest-xdist>=3.5.0
pytest-xprocess==1.0.2
